fixes #262 NNG_OPT_URL should be resolved
This causes TCP, TLS, and ZT endpoints to resolve any
wildcards, and even IP addresses, when reporting the listen
URL.  The dialer URL is reported unresolved.  Test cases
for this are added as well, and nngcat actually reports this
if --verbose is supplied.

+++ src/core/defs.h
@@ -111,8 +111,8 @@ typedef struct {
 	} while (0)
 
 #define NNI_GET16(ptr, v)                            \
-	v = (((uint32_t)((uint8_t)(ptr)[0])) << 8) + \
-	    (((uint32_t)(uint8_t)(ptr)[1]))
+	v = (((uint16_t)((uint8_t)(ptr)[0])) << 8) + \
+	    (((uint16_t)(uint8_t)(ptr)[1]))
 
 #define NNI_GET32(ptr, v)                             \
 	v = (((uint32_t)((uint8_t)(ptr)[0])) << 24) + \
